+++ nengo/builder/processes.py
@@ -65,8 +65,14 @@ class SimProcess(Operator):
         else:
             raise ValueError("Unrecognized mode %r" % mode)
 
-        self.state = {} if state is None else state
-        self.updates.extend(self.state.values())
+        self.state = {}
+        if state is not None:
+            for name, sig in state.items():
+                # The signals actually stored in `self.updates` can be modified by the
+                # optimizer. To allow this possibility, we store the index of the
+                # signal in the updates list instead of storing the signal itself.
+                self.state[name] = len(self.updates)
+                self.updates.append(sig)
 
     @property
     def input(self):
@@ -74,14 +80,12 @@ class SimProcess(Operator):
 
     @property
     def output(self):
-        if len(self.updates) <= len(self.incs) <= len(self.sets) <= 0:
-            return None
-        elif self.mode == "update":
+        if self.mode == "update":
             return self.updates[0]
-        elif self.mode == "inc":
+        if self.mode == "inc":
             return self.incs[0]
-        elif self.mode == "set":
-            return self.sets[0]
+        assert self.mode == "set"
+        return self.sets[0]
 
     @property
     def t(self):
@@ -98,7 +102,7 @@ class SimProcess(Operator):
         shape_in = input.shape if input is not None else (0,)
         shape_out = output.shape
         rng = self.process.get_rng(rng)
-        state = {name: signals[sig] for name, sig in self.state.items()}
+        state = {name: signals[self.updates[idx]] for name, idx in self.state.items()}
         step_f = self.process.make_step(shape_in, shape_out, dt, rng, state)
         args = (t,) if input is None else (t, input)