+++ nengo/neurons.py
@@ -2,11 +2,14 @@ import warnings
 
 import numpy as np
 
+from nengo.dists import Choice, Distribution, get_samples, Uniform
 from nengo.exceptions import SimulationError, ValidationError
-from nengo.params import Parameter, NumberParam, FrozenObject
+from nengo.params import DictParam, FrozenObject, NumberParam, Parameter
+from nengo.rc import rc
+from nengo.utils.numpy import is_array_like
 
 
-def settled_firingrate(step_math, J, states, dt=0.001, settle_time=0.1, sim_time=1.0):
+def settled_firingrate(step, J, state, dt=0.001, settle_time=0.1, sim_time=1.0):
     """Compute firing rates (in Hz) for given vector input, ``x``.
 
     Unlike the default naive implementation, this approach takes into
@@ -17,24 +20,24 @@ def settled_firingrate(step_math, J, states, dt=0.001, settle_time=0.1, sim_time
 
     Parameters
     ----------
-    step_math : function
+    step : function
         the step function of the neuron type
     J : ndarray
         a vector of currents to generate firing rates from
-    *states : list of ndarrays
+    state : dict of ndarrays
         additional state needed by the step function
     """
-    out = np.zeros_like(J)
     total = np.zeros_like(J)
+    out = state.get("spikes", state.get("rates"))
 
     # Simulate for the settle time
     steps = int(settle_time / dt)
     for _ in range(steps):
-        step_math(dt, J, out, *states)
+        step(dt, J, **state)
     # Simulate for sim time, and keep track
     steps = int(sim_time / dt)
     for _ in range(steps):
-        step_math(dt, J, out, *states)
+        step(dt, J, **state)
         total += out
     return total / float(steps)
 
@@ -42,13 +45,60 @@ def settled_firingrate(step_math, J, states, dt=0.001, settle_time=0.1, sim_time
 class NeuronType(FrozenObject):
     """Base class for Nengo neuron models.
 
+    Parameters
+    ----------
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
+
     Attributes
     ----------
-    probeable : tuple
-        Signals that can be probed in the neuron population.
+    state : {str: Distribution}
+        State variables held by the neuron type during simulation.
+        Values in the dict indicate their initial values, or how
+        to obtain those initial values. These elements can also be
+        probed in the neuron population.
     """
 
-    probeable = ()
+    state = {}
+
+    initial_state = DictParam("initial_state", optional=True)
+
+    def __init__(self, initial_state=None):
+        super().__init__()
+        self.initial_state = initial_state
+        if self.initial_state is not None:
+            for name, value in self.initial_state.items():
+                if name not in self.state:
+                    raise ValidationError(
+                        "State variable %r not recognized; should be one of %s"
+                        % (name, ", ".join(repr(k) for k in self.state)),
+                        attr="initial_state",
+                        obj=self,
+                    )
+                if not (isinstance(value, Distribution) or is_array_like(value)):
+                    raise ValidationError(
+                        "State variable %r must be a distribution or array-like"
+                        % (name,),
+                        attr="initial_state",
+                        obj=self,
+                    )
+
+    @property
+    def probeable(self):
+        return tuple(self.state)
+
+    @property
+    def spiking(self):
+        return "spikes" in self.state
+
+    @property
+    def step_math(self):
+        warnings.warn(
+            "'step_math' has been renamed to 'step'. This alias will be removed "
+            "in Nengo 4.0"
+        )
+        return self.step
 
     def current(self, x, gain, bias):
         """Compute current injected in each neuron given input, gain and bias.
@@ -154,6 +204,17 @@ class NeuronType(FrozenObject):
         bias[:] = J_tops - gain
         return gain, bias
 
+    def make_state(self, n_neurons, rng=np.random, dtype=None):
+        dtype = rc.float_dtype if dtype is None else dtype
+        state = {}
+        initial_state = {} if self.initial_state is None else self.initial_state
+        for name in self.state:
+            dist = initial_state.get(name, self.state[name])
+            state[name] = get_samples(dist, n=n_neurons, d=None, rng=rng).astype(
+                dtype, copy=False
+            )
+        return state
+
     def max_rates_intercepts(self, gain, bias):
         """Compute the max_rates and intercepts given gain and bias.
 
@@ -212,10 +273,13 @@ class NeuronType(FrozenObject):
         """
         J = self.current(x, gain, bias)
         out = np.zeros_like(J)
-        self.step_math(dt=1.0, J=J, output=out)
+        if self.spiking:
+            self.step(dt=1.0, J=J, spikes=out)
+        else:
+            self.step(dt=1.0, J=J, rates=out)
         return out
 
-    def step_math(self, dt, J, output):
+    def step(self, dt, J, **state):
         """Implements the differential equation for this neuron type.
 
         At a minimum, NeuronType subclasses must implement this method.
@@ -228,10 +292,10 @@ class NeuronType(FrozenObject):
             Simulation timestep.
         J : (n_neurons,) array_like
             Input currents associated with each neuron.
-        output : (n_neurons,) array_like
-            Output activities associated with each neuron.
+        state : {str: array_like}
+            State variables associated with the population.
         """
-        raise NotImplementedError("Neurons must provide step_math")
+        raise NotImplementedError("Neurons must provide step")
 
 
 class Direct(NeuronType):
@@ -255,18 +319,13 @@ class Direct(NeuronType):
         """Always returns ``x``."""
         return np.array(x, dtype=float, copy=False, ndmin=1)
 
-    def step_math(self, dt, J, output):
+    def step(self, dt, J):
         """Raises an error if called.
 
         Rather than calling this function, the simulator will detect that
         the ensemble is in direct mode, and bypass the neural approximation.
         """
         raise SimulationError("Direct mode neurons shouldn't be simulated.")
-
-
-# TODO: class BasisFunctions or Population or Express;
-#       uses non-neural basis functions to emulate neuron saturation,
-#       but still simulate very fast
 
 
 class RectifiedLinear(NeuronType):
@@ -281,14 +340,17 @@ class RectifiedLinear(NeuronType):
     amplitude : float
         Scaling factor on the neuron output. Corresponds to the relative
         amplitude of the output of the neuron.
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
     """
 
-    probeable = ("rates",)
+    state = {"rates": Choice([0])}
 
     amplitude = NumberParam("amplitude", low=0, low_open=True)
 
-    def __init__(self, amplitude=1):
-        super().__init__()
+    def __init__(self, amplitude=1, initial_state=None):
+        super().__init__(initial_state)
 
         self.amplitude = amplitude
 
@@ -306,9 +368,9 @@ class RectifiedLinear(NeuronType):
         max_rates = gain * (1 - intercepts)
         return max_rates, intercepts
 
-    def step_math(self, dt, J, output):
+    def step(self, dt, J, rates):
         """Implement the rectification nonlinearity."""
-        output[...] = self.amplitude * np.maximum(0.0, J)
+        rates[...] = self.amplitude * np.maximum(0.0, J)
 
 
 class SpikingRectifiedLinear(RectifiedLinear):
@@ -324,24 +386,27 @@ class SpikingRectifiedLinear(RectifiedLinear):
     amplitude : float
         Scaling factor on the neuron output. Corresponds to the relative
         amplitude of the output spikes of the neuron.
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
     """
 
-    probeable = ("spikes", "voltage")
+    state = {"spikes": Choice([0]), "voltage": Uniform(low=0, high=1)}
 
     def rates(self, x, gain, bias):
         """Use RectifiedLinear to determine rates."""
 
         J = self.current(x, gain, bias)
         out = np.zeros_like(J)
-        RectifiedLinear.step_math(self, dt=1.0, J=J, output=out)
+        RectifiedLinear.step(self, dt=1.0, J=J, rates=out)
         return out
 
-    def step_math(self, dt, J, spiked, voltage):
+    def step(self, dt, J, spikes, voltage):
         """Implement the integrate and fire nonlinearity."""
 
         voltage += np.maximum(J, 0) * dt
         n_spikes = np.floor(voltage)
-        spiked[:] = self.amplitude * n_spikes / dt
+        spikes[:] = self.amplitude * n_spikes / dt
         voltage -= n_spikes
 
 
@@ -351,14 +416,20 @@ class Sigmoid(NeuronType):
     Since the tuning curves are strictly positive, the ``intercepts``
     correspond to the inflection point of each sigmoid. That is,
     ``f(intercept) = 0.5`` where ``f`` is the pure sigmoid function.
+
+    Parameters
+    ----------
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
     """
 
-    probeable = ("rates",)
+    state = {"rates": Choice([0])}
 
     tau_ref = NumberParam("tau_ref", low=0)
 
-    def __init__(self, tau_ref=0.0025):
-        super().__init__()
+    def __init__(self, tau_ref=0.0025, initial_state=None):
+        super().__init__(initial_state)
         self.tau_ref = tau_ref
 
     def gain_bias(self, max_rates, intercepts):
@@ -379,9 +450,9 @@ class Sigmoid(NeuronType):
         max_rates = lim / (1 + np.exp(-inverse))
         return max_rates, intercepts
 
-    def step_math(self, dt, J, output):
+    def step(self, dt, J, rates):
         """Implement the sigmoid nonlinearity."""
-        output[...] = (1.0 / self.tau_ref) / (1.0 + np.exp(-J))
+        rates[...] = (1.0 / self.tau_ref) / (1.0 + np.exp(-J))
 
 
 class LIFRate(NeuronType):
@@ -398,16 +469,19 @@ class LIFRate(NeuronType):
     amplitude : float
         Scaling factor on the neuron output. Corresponds to the relative
         amplitude of the output spikes of the neuron.
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
     """
 
-    probeable = ("rates",)
+    state = {"rates": Choice([0])}
 
     tau_rc = NumberParam("tau_rc", low=0, low_open=True)
     tau_ref = NumberParam("tau_ref", low=0)
     amplitude = NumberParam("amplitude", low=0, low_open=True)
 
-    def __init__(self, tau_rc=0.02, tau_ref=0.002, amplitude=1):
-        super().__init__()
+    def __init__(self, tau_rc=0.02, tau_ref=0.002, amplitude=1, initial_state=None):
+        super().__init__(initial_state)
         self.tau_rc = tau_rc
         self.tau_ref = tau_ref
         self.amplitude = amplitude
@@ -448,15 +522,15 @@ class LIFRate(NeuronType):
         """Always use LIFRate to determine rates."""
         J = self.current(x, gain, bias)
         out = np.zeros_like(J)
-        # Use LIFRate's step_math explicitly to ensure rate approximation
-        LIFRate.step_math(self, dt=1, J=J, output=out)
+        # Use LIFRate's step explicitly to ensure rate approximation
+        LIFRate.step(self, dt=1, J=J, rates=out)
         return out
 
-    def step_math(self, dt, J, output):
+    def step(self, dt, J, rates):
         """Implement the LIFRate nonlinearity."""
         j = J - 1
-        output[:] = 0  # faster than output[j <= 0] = 0
-        output[j > 0] = self.amplitude / (
+        rates[:] = 0  # faster than output[j <= 0] = 0
+        rates[j > 0] = self.amplitude / (
             self.tau_ref + self.tau_rc * np.log1p(1.0 / j[j > 0])
         )
         # the above line is designed to throw an error if any j is nan
@@ -480,17 +554,31 @@ class LIF(LIFRate):
     amplitude : float
         Scaling factor on the neuron output. Corresponds to the relative
         amplitude of the output spikes of the neuron.
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
     """
 
-    probeable = ("spikes", "voltage", "refractory_time")
+    state = {
+        "spikes": Choice([0]),
+        "voltage": Uniform(low=0, high=1),
+        "refractory_time": Choice([0]),
+    }
 
     min_voltage = NumberParam("min_voltage", high=0)
 
-    def __init__(self, tau_rc=0.02, tau_ref=0.002, min_voltage=0, amplitude=1):
-        super().__init__(tau_rc=tau_rc, tau_ref=tau_ref, amplitude=amplitude)
+    def __init__(
+        self, tau_rc=0.02, tau_ref=0.002, min_voltage=0, amplitude=1, initial_state=None
+    ):
+        super().__init__(
+            tau_rc=tau_rc,
+            tau_ref=tau_ref,
+            amplitude=amplitude,
+            initial_state=initial_state,
+        )
         self.min_voltage = min_voltage
 
-    def step_math(self, dt, J, spiked, voltage, refractory_time):
+    def step(self, dt, J, spikes, voltage, refractory_time):
         # reduce all refractory times by dt
         refractory_time -= dt
 
@@ -507,7 +595,7 @@ class LIF(LIFRate):
 
         # determine which neurons spiked (set them to 1/dt, else 0)
         spiked_mask = voltage > 1
-        spiked[:] = spiked_mask * (self.amplitude / dt)
+        spikes[:] = spiked_mask * (self.amplitude / dt)
 
         # set v(0) = 1 and solve for t to compute the spike time
         t_spike = dt + self.tau_rc * np.log1p(
@@ -548,6 +636,9 @@ class AdaptiveLIFRate(LIFRate):
     amplitude : float
         Scaling factor on the neuron output. Corresponds to the relative
         amplitude of the output spikes of the neuron.
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
 
     References
     ----------
@@ -556,21 +647,34 @@ class AdaptiveLIFRate(LIFRate):
        16.10 (2004): 2101-2124.
     """
 
-    probeable = ("rates", "adaptation")
+    state = {"rates": Choice([0]), "adaptation": Choice([0])}
 
     tau_n = NumberParam("tau_n", low=0, low_open=True)
     inc_n = NumberParam("inc_n", low=0)
 
-    def __init__(self, tau_n=1, inc_n=0.01, tau_rc=0.02, tau_ref=0.002, amplitude=1):
-        super().__init__(tau_rc=tau_rc, tau_ref=tau_ref, amplitude=amplitude)
+    def __init__(
+        self,
+        tau_n=1,
+        inc_n=0.01,
+        tau_rc=0.02,
+        tau_ref=0.002,
+        amplitude=1,
+        initial_state=None,
+    ):
+        super().__init__(
+            tau_rc=tau_rc,
+            tau_ref=tau_ref,
+            amplitude=amplitude,
+            initial_state=initial_state,
+        )
         self.tau_n = tau_n
         self.inc_n = inc_n
 
-    def step_math(self, dt, J, output, adaptation):
+    def step(self, dt, J, rates, adaptation):
         """Implement the AdaptiveLIFRate nonlinearity."""
         n = adaptation
-        super().step_math(dt, J - n, output)
-        n += (dt / self.tau_n) * (self.inc_n * output - n)
+        super().step(dt, J - n, rates)
+        n += (dt / self.tau_n) * (self.inc_n * rates - n)
 
 
 class AdaptiveLIF(LIF):
@@ -603,6 +707,9 @@ class AdaptiveLIF(LIF):
     amplitude : float
         Scaling factor on the neuron output. Corresponds to the relative
         amplitude of the output spikes of the neuron.
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
 
     References
     ----------
@@ -611,7 +718,12 @@ class AdaptiveLIF(LIF):
        16.10 (2004): 2101-2124.
     """
 
-    probeable = ("spikes", "adaptation", "voltage", "refractory_time")
+    state = {
+        "spikes": Choice([0]),
+        "voltage": Uniform(low=0, high=1),
+        "refractory_time": Choice([0]),
+        "adaptation": Choice([0]),
+    }
 
     tau_n = NumberParam("tau_n", low=0, low_open=True)
     inc_n = NumberParam("inc_n", low=0)
@@ -624,18 +736,23 @@ class AdaptiveLIF(LIF):
         tau_ref=0.002,
         min_voltage=0,
         amplitude=1,
+        initial_state=None,
     ):
         super().__init__(
-            tau_rc=tau_rc, tau_ref=tau_ref, min_voltage=min_voltage, amplitude=amplitude
+            tau_rc=tau_rc,
+            tau_ref=tau_ref,
+            min_voltage=min_voltage,
+            amplitude=amplitude,
+            initial_state=initial_state,
         )
         self.tau_n = tau_n
         self.inc_n = inc_n
 
-    def step_math(self, dt, J, output, voltage, ref, adaptation):
+    def step(self, dt, J, spikes, voltage, refractory_time, adaptation):
         """Implement the AdaptiveLIF nonlinearity."""
         n = adaptation
-        super().step_math(dt, J - n, output, voltage, ref)
-        n += (dt / self.tau_n) * (self.inc_n * output - n)
+        super().step(dt, J - n, spikes, voltage, refractory_time)
+        n += (dt / self.tau_n) * (self.inc_n * spikes - n)
 
 
 class Izhikevich(NeuronType):
@@ -670,6 +787,9 @@ class Izhikevich(NeuronType):
         (Originally 'c') The voltage to reset to after a spike, in millivolts.
     reset_recovery : float, optional
         (Originally 'd') The recovery value to reset to after a spike.
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
 
     References
     ----------
@@ -678,7 +798,11 @@ class Izhikevich(NeuronType):
        (http://www.izhikevich.org/publications/spikes.pdf)
     """
 
-    probeable = ("spikes", "voltage", "recovery")
+    state = {
+        "spikes": Choice([0]),
+        "voltage": Uniform(low=0, high=1),
+        "recovery": Choice([0]),
+    }
 
     tau_recovery = NumberParam("tau_recovery", low=0, low_open=True)
     coupling = NumberParam("coupling", low=0)
@@ -686,9 +810,14 @@ class Izhikevich(NeuronType):
     reset_recovery = NumberParam("reset_recovery")
 
     def __init__(
-        self, tau_recovery=0.02, coupling=0.2, reset_voltage=-65.0, reset_recovery=8.0
+        self,
+        tau_recovery=0.02,
+        coupling=0.2,
+        reset_voltage=-65.0,
+        reset_recovery=8.0,
+        initial_state=None,
     ):
-        super().__init__()
+        super().__init__(initial_state)
         self.tau_recovery = tau_recovery
         self.coupling = coupling
         self.reset_voltage = reset_voltage
@@ -697,13 +826,19 @@ class Izhikevich(NeuronType):
     def rates(self, x, gain, bias):
         """Estimates steady-state firing rate given gain and bias."""
         J = self.current(x, gain, bias)
-        voltage = np.zeros_like(J)
-        recovery = np.zeros_like(J)
         return settled_firingrate(
-            self.step_math, J, [voltage, recovery], settle_time=0.001, sim_time=1.0
+            self.step,
+            J,
+            state={
+                "spikes": np.zeros_like(J),
+                "voltage": np.zeros_like(J),
+                "recovery": np.zeros_like(J),
+            },
+            settle_time=0.001,
+            sim_time=1.0,
         )
 
-    def step_math(self, dt, J, spiked, voltage, recovery):
+    def step(self, dt, J, spikes, voltage, recovery):
         """Implement the Izhikevich nonlinearity."""
         # Numerical instability occurs for very low inputs.
         # We'll clip them be greater than some value that was chosen by
@@ -719,12 +854,12 @@ class Izhikevich(NeuronType):
         # However, calculating recovery for voltage values greater than
         # threshold can cause the system to blow up, which we want
         # to avoid at all costs.
-        spiked[:] = (voltage >= 30) / dt
-        voltage[spiked > 0] = self.reset_voltage
+        spikes[:] = (voltage >= 30) / dt
+        voltage[spikes > 0] = self.reset_voltage
 
         dU = (self.tau_recovery * (self.coupling * voltage - recovery)) * 1000
         recovery[:] += dU * dt
-        recovery[spiked > 0] = recovery[spiked > 0] + self.reset_recovery
+        recovery[spikes > 0] = recovery[spikes > 0] + self.reset_recovery
 
 
 class NeuronTypeParam(Parameter):